refactor(app): remove alert store, move to provide/inject setup, composables
